requirements.txt
django
psycopg2-binary
django-cors-headers
firebase-admin
certifi
python-decouple
djangorestframework-simplejwt